cityListFragment classes moved to cityListFragment package
