+++ app/src/main/java/com/simplertutorials/android/wheathograophy/components/ApplicationComponent.kt
@@ -1,7 +1,7 @@
 package com.simplertutorials.android.wheathograophy.components
 
 import javax.inject.Singleton
-import com.simplertutorials.android.wheathograophy.ui.fragments.CityListFragment
+import com.simplertutorials.android.wheathograophy.ui.fragments.cityListFragment.CityListFragment
 import com.simplertutorials.android.wheathograophy.ui.fragments.WeatherInfoFragment
 import dagger.Component
