+++ app/src/main/java/com/simplertutorials/android/wheathograophy/ui/fragments/cityListFragment/CityListPresenter.kt
@@ -1,4 +1,4 @@
-package com.simplertutorials.android.wheathograophy.ui.fragments
+package com.simplertutorials.android.wheathograophy.ui.fragments.cityListFragment
 
 import com.simplertutorials.android.wheathograophy.data.database.DatabaseRepository
 import com.simplertutorials.android.wheathograophy.domain.City
